requirements.txt
asgiref==3.7.2
Django==5.0.2
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.9.0
redis==4.5.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
djangorestframework-simplejwt==5.3.1
